Feature/deummy scheduling (#13)
* Develop (#11)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* Refactoring/dummy client structure (#6)

* Develop (#4)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* 더미, 클라이언트 및 로그 리팩토링

* Feature/dummy loopback (#8)

* Develop (#7)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* Refactoring/dummy client structure (#6)

* Develop (#4)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* 더미, 클라이언트 및 로그 리팩토링

* Zone 변경시 패킷 보내기 않도록 수정

* Feature/dummy move (#10)

* Develop (#9)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* Refactoring/dummy client structure (#6)

* Develop (#4)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* 더미, 클라이언트 및 로그 리팩토링

* Feature/dummy loopback (#8)

* Develop (#7)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* Refactoring/dummy client structure (#6)

* Develop (#4)

* merge feature/organize

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* test comment

* Feature/changepid to develop (#3)

* merge feature/develop

* Class Dummy and Organize Code

* Add DisConnect/ReConnect

* add changepid

* 더미, 클라이언트 및 로그 리팩토링

* Zone 변경시 패킷 보내기 않도록 수정

* 입장시 서버ID 클라ID 연결 및 Zone 주위 정보 받기

* 개인 이동 확인 및 sendbroadcast 로 온 다른 위치 확인#

* 중복된 ReConnect 선언 지우기

* 일부 필요 없는 코드 제

* 기본 구조를 CDummy 는 Client 의 생성 만
직접적인 Logic 은 Client 에

* CSchedule 추가 스케줄 기본 클래스 생성

* CSchedule 추가 스케줄 기본 클래스 생성

* 더미 스케줄링 을 이용한 채널 테스트 환경 완료

+++ CUtill/CUtill.cpp
@@ -21,15 +21,14 @@ int CUtil::Random(int Min, int Max)
 	return dist(gen);
 }
 
-
-st_Vector3F CUtil::RandomVector2F(int Min, int Max)
+st_Vector3F CUtil::RandomVector2F(float Min, float Max)
 {
-	return st_Vector3F((float)Random(Min,Max), 0, (float)Random(Min, Max));
+	return st_Vector3F(Random(Min,Max), 0, Random(Min, Max));
 }
 
-st_Vector3F CUtil::RandomVector3F(int Min, int Max)
+st_Vector3F CUtil::RandomVector3F(float Min, float Max)
 {
-	return st_Vector3F((float)Random(Min, Max), (float)Random(Min, Max), (float)Random(Min, Max));
+	return st_Vector3F(Random(Min, Max), Random(Min, Max), Random(Min, Max));
 }
 
 double CUtil::GetQPCNowTime()
